Fix previewer not opening issue
