+++ src/com/ichi2/anki/Reviewer.java
@@ -18,39 +18,54 @@
 
 package com.ichi2.anki;
 
+import android.annotation.SuppressLint;
 import android.content.SharedPreferences;
-import android.os.Bundle;
+import android.content.res.Resources;
+import android.util.DisplayMetrics;
+import android.view.Display;
+import android.view.Menu;
+import android.view.MenuItem;
+import android.view.MotionEvent;
+import android.view.View;
+import android.view.ViewConfiguration;
+import android.widget.FrameLayout;
 
-import com.ichi2.anki.reviewer.WhiteboardListener;
 import com.ichi2.async.DeckTask;
 import com.ichi2.libanki.Collection;
 import com.ichi2.widget.WidgetStatus;
 
 import org.json.JSONException;
 
-public class Reviewer extends AbstractFlashcardViewer implements WhiteboardListener {
+public class Reviewer extends AbstractFlashcardViewer {
     private boolean mHasDrawerSwipeConflicts = false;
+    private boolean mShowWhiteboard = true;
+    private boolean mBlackWhiteboard = true;
     
     @Override
     protected void setTitle() {
         try {
             String[] title = mSched.getCol().getDecks().current().getString("name").split("::");
-            AnkiDroidApp.getCompat().setTitle(this, title[title.length - 1], mInvertedColors);
+            AnkiDroidApp.getCompat().setTitle(this, title[title.length - 1], mNightMode);
             super.setTitle(title[title.length - 1]);
         } catch (JSONException e) {
             throw new RuntimeException(e);
         }
-        AnkiDroidApp.getCompat().setSubtitle(this, "", mInvertedColors);
+        AnkiDroidApp.getCompat().setSubtitle(this, "", mNightMode);
     }
 
 
     @Override
     protected void onCollectionLoaded(Collection col) {
-        setWhiteboardListener(this);
         super.onCollectionLoaded(col);
         // Load the first card and start reviewing. Uses the answer card
         // task to load a card, but since we send null
         // as the card to answer, no card will be answered.
+        
+        mPrefWhiteboard = MetaDB.getWhiteboardState(this, getParentDid());
+        if (mPrefWhiteboard) {
+            setWhiteboardEnabledState(true);
+        }
+
         col.getSched().reset();     // Reset schedule incase card had previous been loaded
         DeckTask.launchDeckTask(DeckTask.TASK_TYPE_ANSWER_CARD, mAnswerCardHandler, new DeckTask.TaskData(mSched, null,
                 0));
@@ -58,6 +73,202 @@ public class Reviewer extends AbstractFlashcardViewer implements WhiteboardListe
         // Since we aren't actually answering a card, decrement the rep count
         mSched.setReps(mSched.getReps() - 1);
         disableDrawerSwipeOnConflicts();
+    }
+
+
+
+    @Override
+    public boolean onOptionsItemSelected(MenuItem item) {
+        // The action bar home/up action should open or close the drawer.
+        // ActionBarDrawerToggle will take care of this.
+        if (getDrawerToggle().onOptionsItemSelected(item)) {
+            return true;
+        }
+        switch (item.getItemId()) {
+
+            case android.R.id.home:
+                closeReviewer(RESULT_OK, true);
+                break;
+
+            case R.id.action_undo:
+                undo();
+                break;
+
+            case R.id.action_mark_card:
+                DeckTask.launchDeckTask(DeckTask.TASK_TYPE_MARK_CARD, mMarkCardHandler, new DeckTask.TaskData(mSched,
+                        mCurrentCard, 0));
+                break;
+
+            case R.id.action_replay:
+                playSounds(true);
+                break;
+
+            case R.id.action_edit:
+                return editCard();
+
+            case R.id.action_bury_card:
+                DeckTask.launchDeckTask(DeckTask.TASK_TYPE_DISMISS_NOTE, mDismissCardHandler, new DeckTask.TaskData(
+                        mSched, mCurrentCard, 4));
+                break;
+
+            case R.id.action_bury_note:
+                DeckTask.launchDeckTask(DeckTask.TASK_TYPE_DISMISS_NOTE, mDismissCardHandler, new DeckTask.TaskData(
+                        mSched, mCurrentCard, 0));
+                break;
+
+            case R.id.action_suspend_card:
+                DeckTask.launchDeckTask(DeckTask.TASK_TYPE_DISMISS_NOTE, mDismissCardHandler, new DeckTask.TaskData(
+                        mSched, mCurrentCard, 1));
+                break;
+
+            case R.id.action_suspend_note:
+                DeckTask.launchDeckTask(DeckTask.TASK_TYPE_DISMISS_NOTE, mDismissCardHandler, new DeckTask.TaskData(
+                        mSched, mCurrentCard, 2));
+                break;
+
+            case R.id.action_delete:
+                showDeleteNoteDialog();
+                break;
+
+            case R.id.action_clear_whiteboard:
+                if (mWhiteboard != null) {
+                    mWhiteboard.clear();    
+                }
+                break;
+
+            case R.id.action_hide_whiteboard:
+                // toggle whiteboard visibility
+                setWhiteboardVisibility(!mShowWhiteboard);
+                refreshActionBar();
+                break;
+
+            case R.id.action_enable_whiteboard:
+                // toggle whiteboard enabled state (and show/hide whiteboard item in action bar)
+                mPrefWhiteboard = ! mPrefWhiteboard;
+                setWhiteboardEnabledState(mPrefWhiteboard);
+                setWhiteboardVisibility(mPrefWhiteboard);
+                refreshActionBar();
+                break;
+
+            case R.id.action_search_dictionary:
+                lookUpOrSelectText();
+                break;
+
+            default:
+                return super.onOptionsItemSelected(item);
+        }
+        return true;
+    }
+
+
+    @SuppressLint("NewApi")
+    @Override
+    public boolean onCreateOptionsMenu(Menu menu) {
+        getMenuInflater().inflate(R.menu.reviewer, menu);
+        Resources res = getResources();
+        if (mCurrentCard != null && mCurrentCard.note().hasTag("marked")) {
+            menu.findItem(R.id.action_mark_card).setTitle(R.string.menu_unmark_card).setIcon(R.drawable.ic_menu_marked);
+        } else {
+            menu.findItem(R.id.action_mark_card).setTitle(R.string.menu_mark_card).setIcon(R.drawable.ic_menu_mark);
+        }
+        if (getCol() != null && getCol().undoAvailable()) {
+            menu.findItem(R.id.action_undo).setEnabled(true).setIcon(R.drawable.ic_menu_revert);
+        } else {
+            menu.findItem(R.id.action_undo).setEnabled(false).setIcon(R.drawable.ic_menu_revert_disabled);
+        }
+        if (mPrefWhiteboard) {
+            // Check if we can forceably squeeze in 3 items into the action bar, if not hide "show whiteboard"
+            if (AnkiDroidApp.SDK_VERSION >= 14 &&  !ViewConfiguration.get(this).hasPermanentMenuKey()) {
+                // Android 4.x device with overflow menu in the action bar and small screen can't
+                // support forcing 2 extra items into the action bar
+                Display display = getWindowManager().getDefaultDisplay();
+                DisplayMetrics outMetrics = new DisplayMetrics ();
+                display.getMetrics(outMetrics);
+                float density  = getResources().getDisplayMetrics().density;
+                float dpWidth  = outMetrics.widthPixels / density;
+                if (dpWidth < 360) {
+                    menu.findItem(R.id.action_hide_whiteboard).setShowAsAction(MenuItem.SHOW_AS_ACTION_IF_ROOM);
+                }
+            }
+            // Configure the whiteboard related items in the action bar
+            menu.findItem(R.id.action_enable_whiteboard).setTitle(R.string.disable_whiteboard);
+            menu.findItem(R.id.action_hide_whiteboard).setVisible(true);
+            menu.findItem(R.id.action_clear_whiteboard).setVisible(true);
+            if (mShowWhiteboard) {
+                //menu.findItem(R.id.action_clear_whiteboard).setIcon(R.drawable.ic_whiteboard_clear_enabled);
+                menu.findItem(R.id.action_hide_whiteboard).setIcon(R.drawable.ic_action_whiteboard_enable_light);
+                menu.findItem(R.id.action_hide_whiteboard).setTitle(R.string.hide_whiteboard);
+            } else {
+                //menu.findItem(R.id.action_clear_whiteboard).setIcon(R.drawable.ic_whiteboard_clear_disabled);
+                menu.findItem(R.id.action_hide_whiteboard).setIcon(R.drawable.ic_action_whiteboard_enable_light_disabled);
+                menu.findItem(R.id.action_hide_whiteboard).setTitle(R.string.show_whiteboard);
+            }
+        } else {
+            menu.findItem(R.id.action_enable_whiteboard).setTitle(R.string.enable_whiteboard);
+        }
+        if (AnkiDroidApp.SDK_VERSION < 11 && !mDisableClipboard) {
+            menu.findItem(R.id.action_search_dictionary).setVisible(true).setEnabled(!mShowWhiteboard)
+                    .setTitle(clipboardHasText() ? Lookup.getSearchStringTitle() : res.getString(R.string.menu_select));
+        }
+        return super.onCreateOptionsMenu(menu);
+    }
+
+
+
+
+    /*
+     * Modify the options menu.
+     * Pick the right icons for the whiteboard actions.
+
+     *
+     * @param menu The menu as is.
+     * @return The result of
+     */
+    @Override
+    public boolean onPrepareOptionsMenu(Menu menu) {
+        // On old Androids this gets called each time the menu is opend, on newer ones (=> 3.0) only after it has been
+        // invalidated. We do *that* on purpose when toggeling the night mode.
+        if (mPrefWhiteboard) {
+            if (mShowWhiteboard){
+                if (mNightMode) {
+                    menu.findItem(R.id.action_clear_whiteboard).setIcon(R.drawable.ic_action_cancel_dark);
+                    menu.findItem(R.id.action_hide_whiteboard).setIcon(R.drawable.ic_action_whiteboard_enable_dark);
+                } else {
+                    menu.findItem(R.id.action_clear_whiteboard).setIcon(R.drawable.ic_action_cancel);
+                    menu.findItem(R.id.action_hide_whiteboard).setIcon(R.drawable.ic_action_whiteboard_enable_light);
+
+                }
+            } else {
+                if (mNightMode) {
+                    menu.findItem(R.id.action_clear_whiteboard).setIcon(R.drawable.ic_action_cancel_dark);
+                    menu.findItem(R.id.action_hide_whiteboard).setIcon(R.drawable.ic_action_whiteboard_enable_dark_disabled);
+                } else {
+                    menu.findItem(R.id.action_clear_whiteboard).setIcon(R.drawable.ic_action_cancel);
+                    menu.findItem(R.id.action_hide_whiteboard).setIcon(R.drawable.ic_action_whiteboard_enable_light_disabled);
+                }
+            }
+        }
+
+        return super.onPrepareOptionsMenu(menu);
+    }
+    
+
+    @Override
+    protected SharedPreferences restorePreferences() {
+        super.restorePreferences();
+        SharedPreferences preferences = AnkiDroidApp.getSharedPrefs(getBaseContext());
+        mBlackWhiteboard = preferences.getBoolean("blackWhiteboard", true);
+        return preferences;
+    }
+    
+    @Override
+    public void fillFlashcard() {
+        super.fillFlashcard();
+        if (!sDisplayAnswer) {
+            if (mShowWhiteboard && mWhiteboard != null) {
+                mWhiteboard.clear();
+            }
+        }
     }
 
 
@@ -82,17 +293,57 @@ public class Reviewer extends AbstractFlashcardViewer implements WhiteboardListe
 
 
     @Override
-    public void onShowWhiteboard() {
-        disableDrawerSwipe();
-    }
-
-
-    @Override
-    public void onHideWhiteboard() {
-        if (!mHasDrawerSwipeConflicts) {
-            enableDrawerSwipe();
+    protected void initControls() {
+        super.initControls();
+        if (mPrefWhiteboard) {
+            setWhiteboardVisibility(mShowWhiteboard);
         }
     }
+
+
+    private void setWhiteboardEnabledState(boolean state) {
+        mPrefWhiteboard = state;
+        MetaDB.storeWhiteboardState(this, getParentDid(), state);
+        if (state && mWhiteboard == null) {
+            createWhiteboard();
+        }
+    }
+
+    // Create the whiteboard
+    private void createWhiteboard() {
+        mWhiteboard = new Whiteboard(this, mNightMode, mBlackWhiteboard);
+        FrameLayout.LayoutParams lp2 = new FrameLayout.LayoutParams(
+                android.view.ViewGroup.LayoutParams.FILL_PARENT, android.view.ViewGroup.LayoutParams.FILL_PARENT);
+        mWhiteboard.setLayoutParams(lp2);
+        FrameLayout fl = (FrameLayout) findViewById(R.id.whiteboard);
+        fl.addView(mWhiteboard);
+
+        mWhiteboard.setOnTouchListener(new View.OnTouchListener() {
+            @Override
+            public boolean onTouch(View v, MotionEvent event) {
+                if (mShowWhiteboard) {
+                    return false;
+                }
+                return getGestureDetector().onTouchEvent(event);
+            }
+        });  
+        mWhiteboard.setEnabled(true);
+    }
+
+    // Show or hide the whiteboard
+    private void setWhiteboardVisibility(boolean state) {
+        mShowWhiteboard = state;
+        if (state) {
+            mWhiteboard.setVisibility(View.VISIBLE);
+            disableDrawerSwipe();
+        } else {
+            mWhiteboard.setVisibility(View.GONE);
+            if (!mHasDrawerSwipeConflicts) {
+                enableDrawerSwipe();
+            }
+        }
+    }
+
     
     private void disableDrawerSwipeOnConflicts() {
         SharedPreferences preferences = AnkiDroidApp.getSharedPrefs(getBaseContext());
